Fixed Junit by having stubs. Also renamed the old command classes and added the new add patient classes
