+++ src/main/java/seedu/duke/command/HelpCommand.java
@@ -17,10 +17,10 @@ public class HelpCommand extends Command {
 
     public static final String MESSAGE_USAGE = COMMAND_WORD + ": Shows program usage instructions.\n"
             + "Example: " + COMMAND_WORD;
-    public static final String COMMANDS_MESSAGE_USAGE = AddCommand.MESSAGE_USAGE
-            + "\n" + UpdateCommand.MESSAGE_USAGE
-            + "\n" + DeleteCommand.MESSAGE_USAGE
-            + "\n" + ListCommand.MESSAGE_USAGE
+    public static final String COMMANDS_MESSAGE_USAGE = AddPatientCommand.MESSAGE_USAGE
+            + "\n" + UpdatePatientCommand.MESSAGE_USAGE
+            + "\n" + DeletePatientCommand.MESSAGE_USAGE
+            + "\n" + ListPatientCommand.MESSAGE_USAGE
             + "\n" + HelpCommand.MESSAGE_USAGE;
 
     /**
