+++ src/test/java/seedu/duke/command/ListPatientCommandTest.java
@@ -13,7 +13,7 @@ import java.util.List;
 
 import static org.junit.jupiter.api.Assertions.*;
 
-class ListCommandTest {
+class ListPatientCommandTest {
 
     private static final String EXPECTED_OUTPUT =
             "{[Name]: s;dlskd;l | [Age]: 23 | [Address]: Li | [Contact Number]: 121}" + System.lineSeparator() +
@@ -26,26 +26,16 @@ class ListCommandTest {
 
     @Test
     void testListWithPatients() {
-        Ui ui = null;
-        Storage storage = new Storage();
-        Patient newPatient1 = new Patient("s;dlskd;l", 23, "Li", "121");
-        Patient newPatient2 = new Patient(" ", 15, " ", "15454455");
-        Patient newPatient3 = new Patient("ewuioaiwoe", 33, "Lo", "1989");
-        Patient newPatient4 = new Patient("aeiwae", 13,  "to", " ");
-        Patient newPatient5 = new Patient("LSDs", -1, "Lis", "12");
-        List<Patient> savedPatientList = new ArrayList<>();
-        savedPatientList.add(newPatient1);
-        savedPatientList.add(newPatient2);
-        savedPatientList.add(newPatient3);
-        savedPatientList.add(newPatient4);
-        savedPatientList.add(newPatient5);
-        PatientList patientList = new PatientList(savedPatientList);
-        ListCommand listCommand = new ListCommand();
+        int ui = 1;
+        int storage = 1;
+        int savedPatientList = 1;
+        PatientListStub.createList(savedPatientList);
+        ListCommandStub listCommandStub = new ListCommandStub();
 
         /** Reused from https://coderanch.com/t/587280/java/assertEquals-println **/
         final ByteArrayOutputStream outContent = new ByteArrayOutputStream();
         System.setOut(new PrintStream(outContent));
-        listCommand.execute(ui,storage);
+        listCommandStub.execute(ui,storage);
         assertEquals(EXPECTED_OUTPUT, outContent.toString());
 
 
@@ -54,18 +44,18 @@ class ListCommandTest {
 
     @Test
     void testListWithoutPatients() {
-        Ui ui = null;
-        Storage storage = new Storage();
+        int ui = 1;
+        int storage = 1;
 
-        List<Patient> savedPatientList = new ArrayList<>();
+        int savedPatientList = 2;
 
-        PatientList patientList = new PatientList(savedPatientList);
-        ListCommand listCommand = new ListCommand();
+        PatientListStub.createList(savedPatientList);
+        ListCommandStub listCommandStub = new ListCommandStub();
 
         /** Reused from https://coderanch.com/t/587280/java/assertEquals-println **/
         final ByteArrayOutputStream outContent = new ByteArrayOutputStream();
         System.setOut(new PrintStream(outContent));
-        listCommand.execute(ui,storage);
+        listCommandStub.execute(ui,storage);
         assertEquals(EMPTY, outContent.toString());
